CCS-32: moved all auth pages to an auth folder
