+++ frontend/src/App.tsx
@@ -4,10 +4,10 @@ import {
   Route,
 } from "react-router-dom";
 import Home from "./pages/Home";
-import Login from "./pages/Login";
-import Signup from "./pages/Signup";
+import Login from "./pages/Auth/Login";
+import Signup from "./pages/Auth/Signup";
 import Dashboard from "./pages/Dashboard";
-import ForgotPassword from "./pages/ForgotPassword";
+import ForgotPassword from "./pages/Auth/ForgotPassword";
 
 function App() {
   return (
